@@ -1,16 +1,16 @@
 {
-  "name": "Fms",
-  "version": "1.0.0",
+  "name": "Adsbold",
+  "version": "2.0.0",
   "description": "Facebook management suite",
   "scripts": {
-    "dev": "webpack -w --progress --config webpack.dev.js",
-    "build": "webpack --progress --config webpack.dev.js",
-    "build-prod": "NODE_ENV=production webpack --progress --config webpack.prod.js",
-    "deploy": "NODE_ENV=production webpack --progress --config webpack.prod.js && ./script/deploy",
     "prestart": "npm i",
-    "start": "node server.js"
+    "start": "node server.js",
+    "dev": "webpack -w --progress --config config/webpack/webpack.dev.js",
+    "build": "webpack --progress --config config/webpack/webpack.dev.js",
+    "build-prod": "NODE_ENV=production webpack --config config/webpack/webpack.prod.js",
+    "build-staging": "NODE_ENV=staging webpack --config config/webpack/webpack.staging.js"
   },
-  "author": "Fms team",
+  "author": "Adsbold team",
   "license": "MIT",
   "dependencies": {
     "axios": "^0.9.1",
@@ -18,12 +18,14 @@
     "form-data": "^2.3.1",
     "jquery": "^2.2.1",
     "muicss": "^0.9.27",
-    "react": "^0.14.7",
+    "react": "^16.0.0",
     "react-bootstrap": "^0.31.2",
     "react-bs-notifier": "^4.3.2",
-    "react-dom": "^0.14.7",
-    "react-router": "^2.8.1",
+    "react-dom": "^16.0.0",
+    "react-redux": "^5.0.6",
     "react-router-dom": "^4.2.2",
+    "redux": "^3.7.2",
+    "redux-thunk": "^2.2.0",
     "socket.io-client": "^2.0.3",
     "store": "^2.0.12",
     "url-search-params-polyfill": "^2.0.1",
