@@ -1,27 +1,23 @@
-'use strict';
+import React from 'react';
+import ReactDOM from 'react-dom';
 
-const React = require('react');
-const ReactDOM = require('react-dom');
-
-let FmsToolTip = React.createClass({
-  getDefaultProps: function() {
-    return {
-      message: "",
-      direction: "right"
-    }
-  },
-  componentDidMount: function () {
+class FmsToolTip extends React.Component {
+  componentDidMount() {
     $(ReactDOM.findDOMNode(this.refs.tooltip)).ready(function () {
       $('[data-toggle="tooltip"]').tooltip();
     });
-  },
-  render: function () {
+  }
+  render() {
     return (
       <a ref="tooltip" data-toggle="tooltip" title={this.props.message} data-placement={this.props.direction}>
         {this.props.children}
       </a>
     )
   }
-});
+}
+FmsToolTip.defaultProps = {
+  message: "",
+  direction: "right"
+}
 
 module.exports = FmsToolTip;
