@@ -1,18 +1,17 @@
-'use strict';
 
-const React = require('react');
 
-const notepadImg = require('notepad.png');
+import React from 'react';
+import notepadImg from '../../../images/notepad.png';
 
-let FmsInformationTab = React.createClass({
-  render: function () {
+class FmsInformationTab extends React.Component {
+  render() {
     return (
       <div className="information-tab">
-        <img src={notepadImg} className="note-icon"/>
-        <input type="text" className="note-form" placeholder="Nhập nội dung ghi chú"/>
+        <img src={notepadImg} className="note-icon" />
+        <input type="text" className="note-form" placeholder="Nhập nội dung ghi chú" />
       </div>
     );
   }
-});
+}
 
 module.exports = FmsInformationTab;
