@@ -1,38 +1,32 @@
-'use strict';
 
-const React = require('react');
-const ReactDOM = require('react-dom');
 
-const searchImg = require('search.png');
-let FmsClientItem = require('FmsClientItem');
-let DashboardAPI = require('DashboardApi');
-let FmsSpin = require('FmsSpin');
-let FmsFilterTags = require('FmsFilterTags');
+import React from 'react';
+import ReactDOM from 'react-dom';
 
-let FmsClientList = React.createClass({
-	getInitialState: function() {
-		return {
+import searchImg from '../../../images/search.png';
+import FmsClientItem from './FmsClientItem';
+import DashboardAPI from '../../../api/DashboardApi';
+import FmsSpin from '../../../components/FmsSpin';
+import FmsFilterTags from './FmsFilterTags';
+
+class FmsClientList extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
 			showSpin: false
 		}
-	},
-	getDefaultProps: function () {
-		return {
-			conversations: [],
-			currentConversation: null,
-			allConversations: [],
-			paging: null
-		}
-	},
-	handleClientClick: function(fb_id, type) {
+		this.handleClientClick = this.handleClientClick.bind(this);
+	}
+	handleClientClick(fb_id, type) {
 		this.props.handleClientClick(fb_id, type);
-	},
-	isShowAll: function () {
+	}
+	isShowAll() {
 		let filters = this.props.filters.filter((filter) => {
 			return filter.isActive;
 		});
 		return filters.length == 1 && filters[0].type == 'all';
-	},
-	loadMoreConversations: function () {
+	}
+	loadMoreConversations() {
 		if (this.state.showSpin == true || !this.props.paging) return;
 		if (!this.isShowAll()) return;
 		this.setState({ showSpin: true });
@@ -44,16 +38,16 @@ let FmsClientList = React.createClass({
 			this.setState({ showSpin: false });
 			throw new Error(err);
 		});
-	},
-	componentDidMount: function () {
+	}
+	componentDidMount() {
 		const list = ReactDOM.findDOMNode(this.refs.list);
 		list.addEventListener('scroll', () => {
 			if ($(list).scrollTop() + $(list).innerHeight() >= $(list)[0].scrollHeight - 64) {
 				this.loadMoreConversations();
 			}
 		})
-	},
-	render: function () {
+	}
+	render() {
 		let self = this;
 
 		let renderClients = function () {
@@ -62,7 +56,7 @@ let FmsClientList = React.createClass({
 
 			return conversations.map(conversation => {
 				let isSelected = (self.props.currentConversation && self.props.currentConversation._id == conversation._id);
-				return <FmsClientItem data={conversation} key={conversation._id} handleClientClick={self.handleClientClick} isSelected={isSelected}/>
+				return <FmsClientItem data={conversation} key={conversation._id} handleClientClick={self.handleClientClick} isSelected={isSelected} />
 			});
 		};
 
@@ -71,22 +65,28 @@ let FmsClientList = React.createClass({
 		return (
 			<div className="client-list-wrapper">
 				<div className="search-client">
-					<img src={searchImg} className="search-icon"/>
+					<img src={searchImg} className="search-icon" />
 					<input type="text" className="input-search-client" />
 					<FmsFilterTags tags={this.props.tags} handleFilter={this.props.handleFilter}
-								filters={this.props.filters}/>
+						filters={this.props.filters} />
 				</div>
 				<div ref="list" className="scroll-list">
 					<div>
 						{renderClients()}
 					</div>
 					<div className={"client-list-spin" + showSpin}>
-						<FmsSpin size={27}/>
+						<FmsSpin size={27} />
 					</div>
 				</div>
 			</div>
 		);
 	}
-});
+}
+FmsClientList.defaultProps = {
+	conversations: [],
+	currentConversation: null,
+	allConversations: [],
+	paging: null
+}
 
 module.exports = FmsClientList;
