@@ -1,15 +1,18 @@
-const React = require('react');
-const ReactDOM = require('react-dom');
+'use strict';
 
-let FmsToolTip = require('FmsToolTip');
+import React from 'react';
+import ReactDOM from 'react-dom';
 
-let FmsFilterTags = React.createClass({
-	getInitialState: function () {
-		return {
+import FmsToolTip from '../../../components/FmsToolTip';
+
+class FmsFilterTags extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
 			filters: this.props.filters
 		}
-	},
-	handleTagClick: function (_id) {
+	}
+	handleTagClick(_id) {
 		let newFilters = this.props.filters;
 		newFilters.forEach((filter) => {
 			if (filter.isTag && filter.type == _id) {
@@ -18,8 +21,8 @@ let FmsFilterTags = React.createClass({
 		});
 		this.setState({ filters: newFilters });
 		this.props.handleFilter(newFilters);
-	},
-	renderFilterTag: function () {
+	}
+	renderFilterTag() {
 		if (!this.props.tags) return;
 		return this.props.tags.map((tag) => {
 			let style = { backgroundColor: tag.color };
@@ -30,17 +33,17 @@ let FmsFilterTags = React.createClass({
 				});
 				opacity = (selected.length != 0) ? " selected-filter-tag" : " very-blur";
 			} else opacity = " very-blur";
-			return <FmsToolTip message={tag.name} direction="top" key={tag._id}><div className={"filter-tag" + opacity} 
-						 style={style} onClick={() => { this.handleTagClick(tag._id) }}></div></FmsToolTip>;
+			return <FmsToolTip message={tag.name} direction="top" key={tag._id}><div className={"filter-tag" + opacity}
+				style={style} onClick={() => { this.handleTagClick(tag._id) }}></div></FmsToolTip>;
 		});
-	},
-	render: function () {
+	}
+	render() {
 		return (
 			<div className="filter-tag-bar">
 				{this.renderFilterTag()}
 			</div>
 		)
 	}
-});
+}
 
 module.exports = FmsFilterTags;
