@@ -1,21 +1,17 @@
-const React = require('react');
+import React from 'react';
+import FmsSpin from '../../../components/FmsSpin';
 
-let FmsSpin = require('FmsSpin');
-
-let FmsAttachmentContent = React.createClass({
-  getDefaultProps: function () {
-    return {
-      isSelf: false,
-      hasMessage: -1
-    };
-  },
-  getInitialState: function () {
-    return {
+class FmsAttachmentContent extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {
       size: {},
       isLoading: true
     }
-  },
-  getSizeImage: function () {
+    this.updateChatAreaWidth = this.updateChatAreaWidth.bind(this);
+    this.attachmentLoadDone = this.attachmentLoadDone.bind(this);
+  }
+  getSizeImage() {
     if (!this.props.size || this.props.type == "sticker") return {};
     return {
       width: this.state.size.width,
@@ -23,8 +19,8 @@ let FmsAttachmentContent = React.createClass({
       height: this.state.size.height,
       maxHeight: this.state.size.height
     }
-  },
-  updateChatAreaWidth: function () {
+  }
+  updateChatAreaWidth() {
     if (!this.props.size || this.props.type == "sticker") return;
     let size = {};
     let maxWidth = this.props.getChatAreaWidth() * 0.7;
@@ -41,18 +37,18 @@ let FmsAttachmentContent = React.createClass({
       size.width = 450 * oldWidth / oldHeight;
     }
     this.setState({ size: size });
-  },
-  attachmentLoadDone: function () {
+  }
+  attachmentLoadDone() {
     this.setState({ isLoading: false });
-  },
-  componentWillMount: function () {
+  }
+  componentWillMount() {
     if (this.props.type != "sticker") window.addEventListener("resize", this.updateChatAreaWidth);
     this.updateChatAreaWidth();
-  },
-  componentWillUnmount: function() {
-		if (this.props.type != "sticker") window.removeEventListener("resize", this.updateChatAreaWidth);
-	},
-  render: function () {
+  }
+  componentWillUnmount() {
+    if (this.props.type != "sticker") window.removeEventListener("resize", this.updateChatAreaWidth);
+  }
+  render() {
     let self = this;
     let messageAttachment = (this.props.isSelf) ? "right-message-attachment" : "left-message-attachment";
     let hasMessage;
@@ -74,15 +70,19 @@ let FmsAttachmentContent = React.createClass({
       <div className={"message-attachment-wrapper" + hasMessage + msgAttachWrapper}>
         <div className={messageAttachment + ' ' + isSticker + hasBorder} style={this.getSizeImage()}>
           <div className={"attach-spin" + spin}>
-            <FmsSpin size={27}/>
+            <FmsSpin size={27} />
           </div>
           <a href={self.props.origin} target="_blank">
-            <img className={"image-attachment" + imgAttach} src={preview} onLoad={this.attachmentLoadDone}/>
+            <img className={"image-attachment" + imgAttach} src={preview} onLoad={this.attachmentLoadDone} />
           </a>
         </div>
       </div>
     )
   }
-});
+}
+FmsAttachmentContent.defaultProps = {
+  isSelf: false,
+  hasMessage: -1
+}
 
 module.exports = FmsAttachmentContent;
