@@ -1,36 +1,38 @@
-'use strict';
 
-const React = require('react');
-const ReactDOM = require('react-dom');
-const uuid = require('uuid');
 
-let FmsMessageItem = require('FmsMessageItem');
-let FmsMessageForm = require('FmsMessageForm');
-let FmsInfoChat = require('FmsInfoChat');
-let FmsSpin = require('FmsSpin');
-let DashboardAPI = require('DashboardApi');
-let FmsPostInfoConversation = require('FmsPostInfoConversation');
-let FmsTagsBar = require('FmsTagsBar');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import uuid from 'uuid';
+
+import FmsMessageItem from './FmsMessageItem';
+import FmsMessageForm from './FmsMessageForm';
+import FmsInfoChat from './FmsInfoChat';
+import FmsSpin from '../../../components/FmsSpin';
+import DashboardAPI from '../../../api/DashboardApi';
+import FmsPostInfoConversation from './FmsPostInfoConversation';
+import FmsTagsBar from './FmsTagsBar';
 
 let lastScrollPosition;
 let lastLength = 0;
 
-let FmsConversationArea = React.createClass({
-	getInitialState: function () {
-		return {
+class FmsConversationArea extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
 			showSpin: false,
 			postInfo: null
 		}
-	},
-	clientChanged: function () {
+		this.getChatAreaWidth = this.getChatAreaWidth.bind(this);
+	}
+	clientChanged() {
 		this.setState({ postInfo: null });
 		lastLength = 0;
-	},
-	scrollToBottom: function () {
+	}
+	scrollToBottom() {
 		let list = this.refs.chat_area;
 		list.scrollTop = list.scrollHeight;
-	},
-	loadPostInfo: function () {
+	}
+	loadPostInfo() {
 		let current = this.props.currentConversation;
 		if (!this.state.postInfo && current.type == "comment") {
 			this.setState({ showSpin: true });
@@ -41,20 +43,20 @@ let FmsConversationArea = React.createClass({
 				this.setState({ postInfo: "not found", showSpin: false });
 			});
 		}
-	},
-	getChatAreaWidth: function () {
+	}
+	getChatAreaWidth() {
 		let list = this.refs.chat_area;
 		return list.clientWidth;
-	},
-	componentDidMount: function () {
+	}
+	componentDidMount() {
 		let list = this.refs.chat_area;
 		list.addEventListener('scroll', () => {
 			if ($(list).scrollTop() == 0) {
 				this.loadMoreMessages();
 			}
 		});
-	},
-	loadMoreMessages: function () {
+	}
+	loadMoreMessages() {
 		let current = this.props.currentConversation;
 		if (this.props.isLoading || this.state.showSpin) return;
 		if (current.type == "comment" && current.paging) {
@@ -80,28 +82,27 @@ let FmsConversationArea = React.createClass({
 		} else if (current.type == "comment" && current.parent_fb_id) {
 			this.loadPostInfo();
 		} else if (current.type == "inbox") {
-			let pageInfo = {message: " "};
+			let pageInfo = { message: " " };
 			this.setState({ postInfo: pageInfo });
 		}
-	},
-	componentWillUpdate: function () {
+	}
+	componentWillUpdate() {
 		let list = ReactDOM.findDOMNode(this.refs.chat_area);
 		lastScrollPosition = list.scrollHeight - list.scrollTop;
-	},
-	componentDidUpdate: function (prevProp, prevState) {
+	}
+	componentDidUpdate(prevProp, prevState) {
 		let list = ReactDOM.findDOMNode(this.refs.chat_area);
 		list.scrollTop = list.scrollHeight - lastScrollPosition;
 		if (!this.state.postInfo && list.clientHeight + 12 > list.scrollHeight) {
 			this.loadMoreMessages();
 		}
 		if (this.props.currentConversation.children && this.props.currentConversation.children.length != lastLength) {
-			if (lastLength != 0)	list.scrollTop = list.scrollHeight - lastScrollPosition - 51;
+			if (lastLength != 0) list.scrollTop = list.scrollHeight - lastScrollPosition - 51;
 			lastLength = this.props.currentConversation.children.length;
 		}
-	},
-	render: function () {
+	}
+	render() {
 		let self = this;
-
 		let renderConversation = () => {
 			if (self.props.currentConversation && Array.isArray(self.props.currentConversation.children)) {
 				let messages = self.props.currentConversation.children;
@@ -124,20 +125,20 @@ let FmsConversationArea = React.createClass({
 					let type = (self.props.currentConversation.type == "comment") ? "comment" : "inbox";
 
 					return <FmsMessageItem message={message} key={uuid()} isSelf={isSelf} isLast={isLast} type={type}
-							 getChatAreaWidth={self.getChatAreaWidth} />;
+						getChatAreaWidth={self.getChatAreaWidth} />;
 				});
 			}
 		};
 		let renderPostInfo = () => {
 			if (this.state.postInfo && this.state.postInfo.message) {
-				return <FmsPostInfoConversation content={this.state.postInfo} pageInfo={this.props.currentConversation.page}/>
+				return <FmsPostInfoConversation content={this.state.postInfo} pageInfo={this.props.currentConversation.page} />
 			}
 		};
 		let renderTagsBar = () => {
 			if (this.props.tags && this.props.tags.length > 0 && this.props.isLoading == false)
-					return <FmsTagsBar tags={this.props.tags} conversation={this.props.currentConversation}
-							noti={this.props.noti} updateClientTags={this.props.updateClientTags}
-							alias={this.props.alias}/>
+				return <FmsTagsBar tags={this.props.tags} conversation={this.props.currentConversation}
+					noti={this.props.noti} updateClientTags={this.props.updateClientTags}
+					alias={this.props.alias} />
 		}
 		let showSpin = (this.state.showSpin == true) ? "" : " hide";
 		let chatArea = (this.props.isLoading) ? " hide" : "";
@@ -147,7 +148,7 @@ let FmsConversationArea = React.createClass({
 		return (
 			<div className="inner-conversation-area">
 				<div className="info-chat">
-					<FmsInfoChat currentConversation={this.props.currentConversation} updateBlockCustomer={self.props.updateBlockCustomer}/>
+					<FmsInfoChat currentConversation={this.props.currentConversation} updateBlockCustomer={self.props.updateBlockCustomer} />
 				</div>
 				<div className={"conversation-spin" + spin}>
 					<FmsSpin size={27} />
@@ -161,11 +162,11 @@ let FmsConversationArea = React.createClass({
 				</div>
 				{renderTagsBar()}
 				<div className={"input-message-area" + input}>
-					<FmsMessageForm sendMessage={this.props.sendMessage} conversation={this.props.currentConversation}/>
+					<FmsMessageForm sendMessage={this.props.sendMessage} conversation={this.props.currentConversation} />
 				</div>
 			</div>
 		);
 	}
-});
+}
 
 module.exports = FmsConversationArea;
