@@ -1,8 +1,8 @@
-const React = require('react');
-const uuid = require('uuid');
+import React from 'react';
+import uuid from 'uuid';
 
-let FmsPostInfoConversation = React.createClass({
-  renderAttachments: (attachments) => {
+class FmsPostInfoConversation extends React.Component {
+  renderAttachments(attachments) {
     let atts = [], type = 3;
     if (attachments && Array.isArray(attachments) && attachments.length == 1) {
       let short = attachments[0];
@@ -10,7 +10,7 @@ let FmsPostInfoConversation = React.createClass({
         atts = short.photos;
         type = 1;
       } else if (Array.isArray(short.data) && short.data.length == 1 && short.data[0].subattachments
-                 && Array.isArray(short.data[0].subattachments.data)) {
+        && Array.isArray(short.data[0].subattachments.data)) {
         atts = short.data[0].subattachments.data;
         type = 2;
       }
@@ -20,10 +20,10 @@ let FmsPostInfoConversation = React.createClass({
       if (type == 1) src = att;
       else if (type == 2) src = att.media.image.src;
       return <a className="attachment-in-conversation" href={src} target="_blank" key={uuid()}>
-             <img className="image-in-conversation" src={src}/></a>
+        <img className="image-in-conversation" src={src} /></a>
     });
-  },
-  render: function () {
+  }
+  render() {
     return (
       <div className="post-info-conversation">
         <p className="page-name-conversation">{this.props.pageInfo.name}</p>
@@ -32,6 +32,6 @@ let FmsPostInfoConversation = React.createClass({
       </div>
     )
   }
-});
+}
 
 module.exports = FmsPostInfoConversation;
