@@ -1,24 +1,26 @@
-'use strict';
+import React from 'react';
+import { Modal } from 'react-bootstrap';
 
-const React = require('react');
-const Modal = require('react-bootstrap').Modal;
+import FmsPageItemInModal from '../project/FmsPageItemInModal';
+import PagesAPI from '../../api/PagesApi';
+import FmsSpin from '../components/FmsSpin';
+import * as socket from '../../socket';
 
-let FmsPageItemInModal = require('FmsPageItemInModal');
-let PagesAPI = require('PagesApi');
-let FmsSpin = require('FmsSpin');
-let socket = require('Socket');
-
-let FmsActivePageModal = React.createClass({
-	getInitialState: function () {
-		return {
+class FmsActivePageModal extends React.Component {
+	constructor(props) {
+		super(props);
+		this.state = {
 			isShown: false,
 			selectedPage: null,
 			isDisabled: true,
 			canSelect: true,
 			loadingStatus: null
 		}
-	},
-	handleActiveButton: function () {
+		this.handleActiveButton = this.handleActiveButton.bind(this);
+		this.handleClickOnPageInModal = this.handleClickOnPageInModal.bind(this);
+		this.close = this.close.bind(this);
+	}
+	handleActiveButton() {
 		let self = this;
 		if (!this.state.selectedPage) return;
 
@@ -29,11 +31,11 @@ let FmsActivePageModal = React.createClass({
 				const TIME_DELAY = 300; // miliseconds
 
 				setTimeout(() => {
-					self.setState({loadingStatus : status});
+					self.setState({ loadingStatus: status });
 				}, i * TIME_DELAY);
 			}
 
-			switch(data.type) {
+			switch (data.type) {
 				case 'inbox':
 					let users = data.items;
 
@@ -50,10 +52,10 @@ let FmsActivePageModal = React.createClass({
 						if (!title) return '';
 						return (title.length > 39) ? (title.substring(0, 37) + '...') : title;
 					})
-					.forEach((title, index) => {
-						let _loadingStatus = 'Lấy nội dung bài đăng: ' + title;
-						updateStatus(_loadingStatus, index);
-					});
+						.forEach((title, index) => {
+							let _loadingStatus = 'Lấy nội dung bài đăng: ' + title;
+							updateStatus(_loadingStatus, index);
+						});
 					break;
 			}
 		};
@@ -75,8 +77,8 @@ let FmsActivePageModal = React.createClass({
 		});
 
 		this.setState({ isDisabled: true, canSelect: false });
-	},
-	handleClickOnPageInModal: function (isSelected, page_fb_id) {
+	}
+	handleClickOnPageInModal(isSelected, page_fb_id) {
 		let inactive = this.props.inactive;
 		for (let page of this.props.inactive) {
 			if (page.fb_id == page_fb_id) {
@@ -86,7 +88,7 @@ let FmsActivePageModal = React.createClass({
 				});
 			}
 		}
-	},
+	}
 	renderPagesInModal() {
 		let self = this;
 		let pages = this.props.inactive;
@@ -97,21 +99,21 @@ let FmsActivePageModal = React.createClass({
 			let isSelected = self.state.selectedPage && self.state.selectedPage.fb_id == page.fb_id
 				&& self.state.selectedPage != null;
 			return <FmsPageItemInModal data={page} key={page.fb_id}
-				onPageClick={self.handleClickOnPageInModal} isSelected={isSelected} canSelect={self.state.canSelect}/>
+				onPageClick={self.handleClickOnPageInModal} isSelected={isSelected} canSelect={self.state.canSelect} />
 		});
-	},
-	open: function () {
+	}
+	open() {
 		this.setState({
 			isShown: true,
 			selectedPage: null,
 			isDisabled: true,
 			canSelect: true
 		});
-	},
-	close: function () {
+	}
+	close() {
 		this.setState({ isShown: false });
-	},
-	render: function () {
+	}
+	render() {
 		let loadingStatus = '' + (this.state.loadingStatus || '');
 		let statusHidden = !this.state.canSelect ? ' ' : ' fms-hidden';
 		let canClose = this.state.canSelect;
@@ -138,6 +140,6 @@ let FmsActivePageModal = React.createClass({
 			</Modal>
 		);
 	}
-});
+}
 
 module.exports = FmsActivePageModal;
