@@ -1,54 +1,27 @@
-'use strict';
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
+import { Link } from 'react-router-dom';
 
-import React from 'react';
-import {Link} from 'react-router-dom';
-import FmsProjectItem from 'FmsProjectItem';
-import FmsAddProjectModal from 'FmsAddProjectModal';
-import projectApi from 'ProjectApi';
+import FmsProjectItem from './FmsProjectItem';
+import FmsAddProjectModal from './FmsAddProjectModal';
+import {getProjects, openModal} from '../../actions/project';
 
-let FmsProject = React.createClass({
-  getInitialState: function () {
-    return {
-      projects: null
-    }
-  },
-  componentDidMount: function () {
+class FmsProject extends Component {
+
+  componentDidMount() {
+    const {dispatch} = this.props;
+    dispatch(getProjects());
+  }
+
+  renderPageItems() {
     let self = this;
+    let {projects, dispatch} = this.props;
 
-    this.updateProjects();
-  },
-  updateProjects: function () {
-    let self = this;
-
-    projectApi.getAllProject()
-      .then(projects => {
-        self.setState({projects});
-      })
-      .catch(err => {
-        alert(err.message);
-      })
-  },
-  handleDeleteProject: function (alias) {
-    let self = this;
-
-    projectApi.deleteProject(alias)
-      .then(() => {
-        self.updateProjects();
-      })
-      .catch(err => {
-        alert(err);
-      })
-  },
-  renderPageItems: function () {
-    let self = this;
-    let projects = this.state.projects;
-
-    if (projects && Array.isArray(projects) && projects.length > 0) {
+    if (projects.length > 0) {
       return projects.map(project => {
         return (
           <Link key={project.alias} to={self.props.match.path + '/' + project.alias}>
-            <FmsProjectItem data={project}
-              handleDeleteProject={self.handleDeleteProject}></FmsProjectItem>
+            <FmsProjectItem data={project}/>
           </Link>
         )
       })
@@ -57,32 +30,36 @@ let FmsProject = React.createClass({
         <div>Bạn chưa có project nào</div>
       )
     }
-  },
-  openModal: function () {
-    this._child.open();
-  },
-  render: function() {
-    let self = this;
+  }
 
+  openModal() {
+    let { dispatch } = this.props;
+    dispatch(openModal());
+  }
+
+  render() {
     return (
       <div className="page container">
         <div className="project-wrapper">
           <div className="row button-project-wrapper">
             <div className="col-md-2">
-              <button className="btn btn-primary" onClick={self.openModal}>Add new project</button>
+              <button className="btn btn-primary" onClick={this.openModal.bind(this)}>Add new project</button>
             </div>
           </div>
-
           <div className="row">
-            {self.renderPageItems()}
+            { this.renderPageItems() }
           </div>
-
-          <FmsAddProjectModal ref={(child) => {this._child = child;}}
-            updateProjects={self.updateProjects} {...self.props}></FmsAddProjectModal>
+          <FmsAddProjectModal />
         </div>
       </div>
     );
   }
-});
+}
 
-module.exports = FmsProject;
+const mapStateToProps = state => {
+  return {
+    projects: state.project.projects
+  }
+}
+
+export default connect(mapStateToProps)(FmsProject);
