@@ -1,29 +1,24 @@
-'use strict';
+import React from 'react';
+import uuid from 'uuid';
+import {connect} from 'react-redux';
 
-const React = require('react');
-const {browserHistory} = require('react-router');
-const uuid = require('uuid');
+import projectApi from '../../api/ProjectApi';
+import { deleteProject } from '../../actions/project';
 
-import projectApi from 'ProjectApi';
-
-let FmsProjectItem = React.createClass({
-  renderPageItem: function () {
-    let self = this;
+class FmsProjectItem extends React.Component {
+  renderPageItem() {
     let project = this.props.data;
     let pages = project.pages;
     const MAX_ITEM = 5;
-
     if (pages && Array.isArray(pages) && pages.length > 0) {
       let pageComponents = pages.filter((item, index) => {
         return index <= MAX_ITEM;
-      })
-      .map(page => {
-        let pageAva = `https://graph.facebook.com/v2.10/${page.fb_id}/picture`;
-
-        return (
-          <img key={page.fb_id} src={pageAva}></img>
-        )
-      });
+      }).map(page => {
+          let pageAva = `https://graph.facebook.com/v2.10/${page.fb_id}/picture`;
+          return (
+            <img key={page.fb_id} src={pageAva}></img>
+          )
+        });
 
       if (pages.length > MAX_ITEM) {
         let moreText = '+' + (pages.length - MAX_ITEM);
@@ -34,17 +29,15 @@ let FmsProjectItem = React.createClass({
     } else {
       return <div></div>
     }
-  },
-  deleteProjectClick: function (e) {
+  }
+  deleteProjectClick(e) {
     e.preventDefault();
     e.stopPropagation();
-    let self = this;
-    let projectAlias = this.props.data.alias;
-
-    self.props.handleDeleteProject(projectAlias);
-  },
-  render: function() {
-    let self = this;
+    let { data, dispatch } = this.props;
+    let projectAlias = data.alias;
+    dispatch(deleteProject(projectAlias));
+  }
+  render() {
     let project = this.props.data;
     let projectName = project.name;
 
@@ -56,13 +49,16 @@ let FmsProjectItem = React.createClass({
             <span className="glyphicons glyphicons-bin"></span>
           </div>
           <div className="panel-body">
-            <div className="page-wrapper">{self.renderPageItem()}</div>
-            <button className="btn btn-danger" onClick={self.deleteProjectClick}>Delete</button>
+            <div className="page-wrapper">{this.renderPageItem()}</div>
+            <button className="btn btn-danger" onClick={this.deleteProjectClick.bind(this)}>Delete</button>
           </div>
         </div>
       </div>
     );
   }
-});
+}
 
-module.exports = FmsProjectItem;
+const mapStateToProps = state => {
+  return {}
+}
+export default connect(mapStateToProps)(FmsProjectItem);
