@@ -1,9 +1,7 @@
-'use strict';
-
-let apiSender = require('ApiSender');
+let apiSender = require('./ApiSender');
 
 module.exports = {
-	getAllProject: (name) => {
+	getAllProjects: () => {
 		let route = `/api/projects`;
 		return apiSender.get(route);
 	},
