@@ -1,61 +1,45 @@
-'use strict';
-
-import React from 'react';
+import React, {Component} from 'react';
+import {connect} from 'react-redux';
 import { AlertList, Alert, AlertContainer } from "react-bs-notifier";
-import {Switch, Route, Redirect} from 'react-router-dom';
-import store from 'store';
+import {Switch, Route, Redirect, withRouter} from 'react-router-dom';
+import * as store from '../helpers/storage';
 import uuid from 'uuid';
+import propTypes from 'prop-types';
 
-import FmsHome from 'FmsHome';
-import FmsProject from 'FmsProject';
-import FmsDashboard from 'FmsDashboard';
-import FmsLogin from 'FmsLogin';
-import FmsNavigation from 'FmsNavigation';
-import FmsPosts from 'FmsPosts';
-import FmsSettings from 'FmsSettings';
-import FmsLoading from 'FmsLoading';
-import FmsRoute from 'FmsRoute';
-import FmsAuthen from 'FmsAuthen';
-import tokenApi from 'TokenApi';
-import socket from 'Socket';
+import FmsHome from './home/FmsHome';
+import FmsProject from './project/FmsProject';
+import FmsDashboard from './dashboard/FmsDashboard';
+import FmsNavigation from '../components/FmsNavigation';
+import FmsPosts from './posts/FmsPosts';
+import FmsSettings from './settings/FmsSettings';
+import FmsLoading from '../components/FmsLoading';
+import FmsRoute from '../components/FmsRoute';
+import tokenApi from '../api/TokenApi';
+import * as socket from '../socket';
+import {ALERT_TIME_DISMIS} from '../constants/utils';
+import {logIn, verifyAccessToken} from '../actions/auth';
 
-import {ALERT_TIME_DISMIS} from 'constant';
-import config from 'CONFIG';
+class FmsApp extends Component {
+  constructor(props) {
+    super(props);
 
-let FmsApp = React.createClass({
-  getInitialState: function () {
-    return {
+    this.state = {
       alerts: [],
       isLoading: true
-    }
-  },
-  componentDidMount: function () {
-    let self = this;
-    // self.noti('success', 'Bỏ ẩn bình luận thành công');
-    // info, warning, danger, or success.
+    };
+  }
 
-    // verify access token
-    let jwt = store.get('jwt');
-    if (jwt) {
-      tokenApi.verifyAccessToken(jwt)
-        .then(userData => {
-          socket.connect(config.BASE_URL, jwt);
-          FmsAuthen.isAuthenticated = true;
-        })
-        .catch(err => {
-          FmsAuthen.clean();
-          FmsAuthen.isAuthenticated = false;
-        })
-        .then(() => {
-          self.setState({ isLoading: false });
-        })
-    } else {
-      FmsAuthen.clean();
-      FmsAuthen.isAuthenticated = false;
-      self.setState({ isLoading: false });
-    }
-  },
-  noti: function (type, message) {
+  componentDidMount() {
+    const {dispatch} = this.props;
+
+    const search = this.props.location.search;
+    const params = new URLSearchParams(search);
+    const access_token = params.get('access_token');
+
+    dispatch(verifyAccessToken(access_token));
+  }
+
+  noti(type, message) {
     let self = this;
 
     let alert = {
@@ -66,69 +50,84 @@ let FmsApp = React.createClass({
 
     let alerts = self.state.alerts;
     alerts.push(alert);
-    self.setState({alerts: alerts});
+    self.setState({ alerts: alerts });
 
     setTimeout(() => {
       self.removeNoti(alert.id);
     }, ALERT_TIME_DISMIS);
-  },
-  removeNoti: function (a_id) {
+  }
+
+  removeNoti(a_id) {
     let self = this;
 
     let alerts = self.state.alerts;
     let filterAlerts = alerts.filter(a => a.id != a_id);
 
     self.setState({alerts: filterAlerts});
-  },
-  renderAlerts: function () {
+  }
+
+  renderAlerts() {
     let self = this;
     let alerts = self.state.alerts;
     let alertItems = alerts.map(alert => {
       return (
-        <Alert key={alert.id} type={alert.type} onDismiss={() => {self.removeNoti(alert.id)}}>{alert.message}</Alert>
+        <Alert key={alert.id} type={alert.type} onDismiss={() => { self.removeNoti(alert.id) }}>{alert.message}</Alert>
       )
     })
 
     return (
       <AlertContainer>
-    		{ alertItems }
-    	</AlertContainer>
+        {alertItems}
+      </AlertContainer>
     )
-  },
-  render: function () {
-    let self = this;
+  }
 
-    if (self.state.isLoading) {
+  render() {
+    const {isLoading, isAuthenticated} = this.props;
+
+    if (isLoading) {
       return (
-        <FmsLoading/>
+        <FmsLoading />
       )
     } else {
-      if (FmsAuthen.isAuthenticated) {
+      if (isAuthenticated) {
         return (
           <div>
-            {self.renderAlerts()}
+            {this.renderAlerts()}
             <FmsNavigation />
             <Switch>
-              <FmsRoute exact path="/" component={FmsHome} noti={self.noti}/>
-              <FmsRoute exact path="/projects" component={FmsProject} noti={self.noti} />
-              <FmsRoute exact path="/projects/:project_alias" component={FmsDashboard} noti={self.noti} />
-              <FmsRoute path="/projects/:project_alias/posts" component={FmsPosts} noti={self.noti} />
-              <FmsRoute path="/projects/:project_alias/settings" component={FmsSettings} noti={self.noti} />
-              <FmsRoute path="/login" component={FmsLogin} noti={self.noti}/>
+              <FmsRoute exact path="/projects" component={FmsProject} noti={this.noti.bind(this)} />
+              <FmsRoute exact path="/projects/:project_alias" component={FmsDashboard} noti={this.noti.bind(this)} />
+              <FmsRoute path="/projects/:project_alias/posts" component={FmsPosts} noti={this.noti.bind(this)} />
+              <FmsRoute path="/projects/:project_alias/settings" component={FmsSettings} noti={this.noti.bind(this)} />
+              <Redirect to="/projects"/>
             </Switch>
           </div>
         )
       } else {
         return (
           <Switch>
-            <FmsRoute exact path="/" component={FmsHome} noti={self.noti}/>
-            <FmsRoute path="/login" component={FmsLogin} noti={self.noti}/>
+            <FmsRoute exact path="/" component={FmsHome} noti={this.noti.bind(this)}/>
             <Redirect to="/"/>
           </Switch>
         )
       }
     }
   }
-});
+}
 
-module.exports = FmsApp;
+FmsApp.propTypes = {
+  isLoading: propTypes.bool.isRequired,
+  isAuthenticated: propTypes.bool.isRequired,
+  user: propTypes.object
+}
+
+const mapStateToProps = (state) => {
+  return {
+    isLoading: state.auth.isLoading,
+    isAuthenticated: state.auth.isAuthenticated,
+    user: state.auth.user
+  }
+}
+
+export default withRouter(connect(mapStateToProps)(FmsApp));
