@@ -1,4 +1,4 @@
-'use strict';
+
 
 const express = require('express');
 const path = require('path');
